requirements.txt
Flask
flask-cors
torch
numpy
opencv-python
qrcode
Pillow
